populate wallets and accounts in one dispatch; fix default wallet selection
