+++ src/account-manager.js
@@ -1,7 +1,6 @@
 import { bindActionCreators } from '/libraries/redux/src/index.js';
 import {
     addAccount,
-    setAllAccounts,
     updateLabel as updateAccountLabel,
     logoutLegacy
 } from '/elements/x-accounts/accounts-redux.js';
@@ -9,7 +8,6 @@ import MixinRedux from '/secure-elements/mixin-redux/mixin-redux.js';
 import AccountsClient from './AccountsClient.standalone.es.js';
 import {
     WalletType,
-    setAllWallets,
     login,
     logout,
     updateLabel as updateWalletLabel,
@@ -17,6 +15,7 @@ import {
     switchWallet,
     setFileFlag,
     setWordsFlag,
+    populate,
     LEGACY
 } from './wallet-redux.js';
 import AccountType from './lib/account-type.js';
@@ -32,7 +31,6 @@ class AccountManager {
 
     constructor() {
         this._launched = new Promise(res => this._resolveLaunched = res);
-        this.accountsLoaded = new Promise(res => this._resolveAccountsLoaded = res);
     }
 
     async launch() {
@@ -187,9 +185,7 @@ class AccountManager {
 
         this.actions = bindActionCreators({
             addAccount,
-            setAllAccounts,
             updateAccountLabel,
-            setAllWallets,
             setDefaultWallet,
             login,
             logout,
@@ -197,7 +193,8 @@ class AccountManager {
             updateWalletLabel,
             switchWallet,
             setFileFlag,
-            setWordsFlag
+            setWordsFlag,
+            populate
         }, this.store.dispatch);
     }
 
@@ -248,48 +245,7 @@ class AccountManager {
             else throw error;
         }
 
-        const wallets = [];
-        const accounts = [];
-
-        listedWallets.forEach(wallet => {
-
-            if (wallet.type !== WalletType.LEGACY) {
-                wallets.push({
-                    id: wallet.accountId,
-                    label: wallet.label,
-                    type: wallet.type,
-                    fileExported: wallet.fileExported,
-                    wordsExported: wallet.wordsExported,
-                });
-            }
-
-            wallet.addresses.forEach(address => {
-                const entry = {
-                    address: address.address,
-                    label: address.label,
-                    type: AccountType.KEYGUARD_HIGH,
-                    isLegacy: wallet.type === WalletType.LEGACY,
-                    walletId: wallet.accountId,
-                };
-                accounts.push(entry);
-            });
-
-            wallet.contracts.forEach(contract => {
-                const entry = Object.assign({}, contract, {
-                    type: AccountType.VESTING,
-                    stepAmount: contract.stepAmount / 1e5,
-                    totalAmount: contract.totalAmount / 1e5,
-                    isLegacy: wallet.type === WalletType.LEGACY,
-                    walletId: wallet.accountId,
-                });
-                accounts.push(entry);
-            });
-        });
-
-        this.actions.setAllAccounts(accounts);
-        this.actions.setAllWallets(wallets);
-
-        this._resolveAccountsLoaded();
+        this.actions.populate(listedWallets);
     }
 
     _onOnboardingResult(result) {
